Adds in a stub test framework
Abstracts a very basic Ginkgo-based test framework and test suite
inspired by https://github.com/openshift-kni/cnf-features-deploy.  For
now, the test suite should always pass, and serves as a stub to add
actual tests in the future.

Additionally, a Makefile is added to accelerate building and running
the code.  In order to build and invoke the tests, just run:

> make

This will build the default target functests, which builds and runs the
test suite.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>