Transition to goexpect from TCL
goexpect is a library that provides a GoLang equivalent to the popular TCL
based expect.  Although subprocessing to TCL allows us to depend on a proven,
trusted tool, it was found that there are a number of differences between the
TCL and GoLang languages that made this a naive proposition:
1) GoLang Regular Expressions differ from TCL ones.
  - TCL utilizes reluctant quantifiers by default, whereas GoLang utilizes
    greedy quantifiers.  Even if this is a setting we can change, there are
    other differences that make this a non-option.
  - TCL does not allow certain regular expression lookbacks, which were deemed
    too expensive to perform.
  Because of these differences, we would likely have had to limit the allowable
  regular expressions for reel.exp.  The other option would be to maintain two
  copies, specially tailored for the particular language at hand.
2) TCL JSON serialization emits some characters that cause GoLang to choke.
   TCL was found to pass certain ANSI characters in JSON payloads which could
   not easily be Unmarshaled in GoLang.
For these reasons, among others, we decided that inclusion of a second language
in the solution is non-scalable.  Thus, goexpect was introduced to replace the
reel.exp implementation.

The goexpect Batcher allows for pretty much anything and everything, which is
really quite undesirable from a testing standpoint.  The existing reel.Reel
finite state machine (FSM) is a much more reliable mechanism, as it ensures
that some expected state is reachable.  Thus, it was decided to wrap gut the
FSM implementation and swap in goexpect to take its place.

An "interactive" directory is abstracted to handle the implementation of
various interactive shells, including:
1) The shell defined by $SHELL
2) ssh
3) oc exec -it ...

These interactive implementations are thoroughly unit tested, and any missed
lines are by virtue of the fact that multithreaded unit testing is difficult.

A "handlers" directory is abstracted to hold all of the existing tests, as well
as a new one called "hostname".  These implementations have 100% line coverage.

In the process of adding unit tests, several mock implementations were added.
An expect.Expecter mock was generated using the mockgen utility, and will need
to be updated whenever we upgrade goexpect.  The rest of the mocks are easily
updated using `make mocks`, which will generate the updated mocks where they
belong.

The `feeder` abstraction was retired, as it did not fit the current needs of
the code base.  It is expected that a similar sort of mechanism may be
required in the future.  Instances of the feeder (i.e., in the cmd/*
directory), have been replaced with more basic illustrations.  This is
considered work in progress, but shouldn't block us from merging this patch.

Lastly, a `unittests` Makefile target was created to run the unit tests with
coverage.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>