CNF Certification GoLang code should be formatted appropriately
The GoLang specification is exacting in terms of its code formatting and
stylistic guidelines.  Whether one loves or hates this dictum, it should be
obeyed in order to promote hygienic GoLang code which adheres to the language
specification.

This change includes a change to the Makefile which automates GoLang code
formatting across the project at build time using the "gofmt" tool.
Additionally, this change includes an initial run of the "gofmt" tool, which
fixes existing formatting errors in the project.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>
Issue-Id: CTONET-508