Define declarative tnf.Test(s) condition framework changes
Defines the "condition" interface, which is "Evaluate"-able.  This is a small
part of a much bigger story defined by CTONET-525, and was chosen first since
it is a leaf node in the dependency tree.

This PR also introduces a number of common string and int conditions.  This
PR has 100% unit test code coverage.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>