Define declarative tnf.Test(s) assertion framework changes
The assertion framework allows specification of an arbitrary number of
"composedAssertion"s.  Each "composedAssertion" can have an arbitrary number of
"assertion"s as well as a designated "logic" type ("and" and "or" are supported
out of the box, though others are possible).

All "composedAssertion"s must evaluate as true at the top level.  Essentially,
a "composedAssertion" is just a bucket, and all buckets must ultimately evaluate
as true.

The decision to create a single module (and resulting super long commit) is
because breaking into separate modules (such as "logic" and "assertion") causes
cyclic dependency issues.

Testing is comprehensive, and uses real world examples.  100% unit test could
not be achieved, as there are some spots in the code that should never get
actually be called.  For example, there are spots where a json.Unmarshal call
to a sub-struct will never fail given the entrypoint.  Instead of ignoring the
errors in these cases "that should never be called", which are programmers'
famous last words, I kept the conditionals.  This is because:
1) The checks are cheap, and may be elided/pruned by the compiler anyway.
2) If the code changes, and these checks become necessary, I don't want to have
   to read the entire code base to determine where I should add them.  This
   is very likely to happen if we decide to implement custom UnmarshalJSON()
   functions for the various types of Conditions.
The lack of testing in the few select areas would likely be caught in an
earlier pass by the JSON Schema linter validation anyway.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>