Introduce multiple container logic to TestConfiguration
This commit contains logic to enable testing across a multitude of containers
from the containerUnderTest and partnerContainer perspective.  The former is
made necessary by the fact that most CNFs are composed of multiple containers.
The latter is future proofing for more complex test cases.

The default test orchestrator is also abstracted.  This is the orchestrator
which will be the "test partner".

All keys are complex ContainerIdentifiers.  This provided the most efficient
and human readable storage mechanism.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>