Implement GoLang Linter to increase code quality (#12)
The test-network-function code base is approaching the size that implementing
code quality controls should be mandatory (in hindsight, we ought to have done
this much from the start).  Thus, this patch implements go lint on non-vendor
files.

We have little control over what our dependencies provide in terms of linting;
it would be great to contribute and help them with achieving lint status, but
for now we will simply ignore lint errors caused by our dependencies.  This
patch considers only the code implemented directly in this repository.

As such, a Makefile target called "lint" was introduced, which invokes the
linter on non-vendor files.  Additionally, the default Makefile target (build)
is modified to invoke the "lint" target, so that lint errors are found earlier
on in the process.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>