Define simple tnf.Test(s) declaratively (#30)
* Define simple tnf.Test(s) declaratively

This PR is just the schema for CTONET-525.  The corresponding Go facade
implementation is implemented, but requires further review before it is pushed.

This schema defines a way to create a "Test" locally.  This schema contains
enough information to:
* Issue an initial command through "arguments".
* Define a "step" which is returned by ReelFirst() in the Go facade.
* Make determinations on match of whether a Test passes/fails based on:
- the matched pattern
- combinations assertions on the underlying match groups

Importantly, the REEL Finite State Machine is maintained.  This is done by
allowing "resultContext" objects to define a "nextStep" and "nextResultContext"
array.  Recursion is completely possible, and encouraged!

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>

* fixup! Define simple tnf.Test(s) declaratively