Define declarative tnf.Test(s) resultContext definition (#34)
ResultContext is a type that allows definition of basic ComposedAssertion(s)
over a Match.  ResultContext is recursive;  if a test writer desires, he or she
can define subsequent reel.Step and corresponding generic.ResultContexts for
further calls to ReelMatch.

Due to the fact that the default MarshalJSON included with GoLang returns
"null" for an empty slice of *ResultContext, this custom MarshalJSON
implementation is provided to elide it altogether instead.  This is due to
the fact that "null" will not Unmarshal cleanly.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>