Append new line character for reel.Step.Execute when missing
If a reel.Step.Execute is missing a newline at the end, it is the same as
inputting a command to a terminal and never pressing return.  Thus, this
makes the state machine implementation more defensive by adding a newline
if one is missing.  There is no known use case at this time when this is
not desired.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>