change "dupl" linter config and address legitimate duplication
dupl does not search for duplicate code by looking for repeated
strings in source code. It finds potential duplicate code by
finding identical structures in the code, taking into account the
types on which things operate. This is far more effective than
text-based analysis but it also more prone to false positives,
particularly when configured with a small number of tokens being
enough to trigger a match.

For example, with the previous value of 20 tokens, dupl was
detecting a function header in an `interface` as a duplication
of the function header from the implementation. There were also
numerous duplications reported from the data structures used in
test setup, test methods, and handlers for common events that
would have made the code less readable to address.

After examining all the issues highlighted with the minimum
token sequence size at 20, I decided to address one issue by
changing the code, and increase the sequence size to allow the
others to pass. The new value of 75 is still smaller than most
and may need to be raised again in the future if it becomes
problematic.

Signed-off-by: Charlie Wheeler-Robinson <cwheeler@redhat.com>