update golangci-lint config with corrections and improvements
"nolintlint" now requires machine-readable formatting: nolint
comments with a space ("// nolint") appear in `godoc` output
while ones without a space ("//nolint") do not. This is important
when disabling linting for more than a single line, e.g. for the
scope of a single function or entire file, where the nolint comment
is placed with the document comment.

Wrapping long comments is desirable so the exception for them in
the linter config is removed. The exception was also misleading;
as written, only empty comment lines ("// ") were being ignored.

Comments in the config have also been adjusted to better match
the config.

Signed-off-by: Charlie Wheeler-Robinson <cwheeler@redhat.com>