Introduces an identifier definition for tnf.Test implementations
Part of the criteria of CTONET-533 and CTONET-541 is to establish an
identification system for tnf.Test implementations.  As such, an identifier
framework is added to the code base.  The tnf.Tester interface was modified
to add a `GetIdentifier()` contract, which returns the unique test identifier.

As such, all existing tnf.Tester implementations were modified to include the
newly added method.  In addition to implementing the method, test code was
added to ensure the identifier is set as expected.

A `Catalog` is introduced to store a map of Identifier to Description/Purpose
and test Result type.  For this iteration of code, we only have "normative"
tests.

Lastly, a CLI is introduced to generate a JSON formatted mapping of the test
catalog, which can be used in documentation later.

A few linter rules were changed:
1) dupl:  This rule erroneously raises duplicate errors.  Sometimes, repeating
   a few lines of code, however similar, is preferable from a readability
   standpoint.  This is much easier than, say, passing function pointers, which
   would otherwise be required.
2) line length:  Effective Go prefers long lines over breaking up a line
   arbitrarily.  In this case, a description exceeded the 200 character limit.
   In today's age, I am unconvinced that mandating strict max line lengths is
   that useful.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>