Introduce jsontest-cli utility
Upon completing CTONET-543, A test developer's guide to writing tests, it
quickly became obvious that the barrier to entry for writing a JSON test is too
high.  Even if one were able to understand the JSON Schema presented in
generic-test.json.schema, he or she may have trouble understanding how to run a
given test, even for just local debugging.  This commit introduces a new
utility that lowers the entrance criteria for creating JSON tests.  This is
then used in the documentation of CTONET-543 to describe local test dev.

`jsontest-cli` is a binary that can be created using `make jsontest-cli`.  The
CLI has a single sub-command for now, `run`.  In the future, we could consider
including `generate` or `validate` commands, but those are extraneous for now.

`run` in turn takes a subcommand;  `oc`, `ssh`, or `shell`.  These three
implementations include the logic to run an arbitrary JSON test in their
specific context.  For example, `oc` requires `namespace`, `pod`, and
`container` arguments.  These are used to set up an interactive shell that can
be used to run the input JSON file.

This CLI is not unit tested, as it is more a target of integration level tests,
which we do not include yet.  It has been tested manually extensively, but
will require future testing work when we include an integration suite.

Documentation for this CLI will be included in a subsequent patch, in the
DEVELOPING.md guide.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>