JSON test developer's guide
This patch includes JSON based test developer's guide using `ping` as an
example.  The guide starts out with how you may manually use `ping` to test
connectivity on the CLI.  It then introduces the JSON defined test for ping,
and explains each facet of the JSON file.  Lastly, it introduces the ability
to run the test using jsontest-cli.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>