Register container and operator configurations with the pool
The container/operator configurations were never registered after
collapsing into a single suite.  This change registers the configs
so they are output as a part of claim.json.

Additionally, it was noticed that the format directive of the
startTime and endTime was not compliant with ISO 8601.  This was
changed appropriately.

Lastly, the claim startTime and endTime are converted to UTC for
consistency.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>