Introduce a "diagnostic" test spec
The diagnostic test suite fulfills capturing a profile of the hardware under
test as according to CTONET-528.  This PR contains the implementation only;
documentation will be added following the close of CTONET-669.  CTONET-669
heavily augments the documentation, and I do not want to introduce a
non-trivial merge conflict;  so I will wait until that is done.

The "diagnostic" test suite is simple;  it just runs the following command:
`oc get nodes -o json`
Due to the fact that we recommend deploying worker nodes on top of CoreOS,
and CoreOS does not have `lspci` or `lshw` baked in, we cannot get any of this
information.  `oc get nodes -o json` does provide us enough information to get
an idea the cluster hardware, and is sufficient for CTONET-528.

As a part of this exercise, I decided to use a JSON based test called
`nodes.json`.  I quickly discovered that a few schema changes were needed in
order to enforce mandatory test identifiers.  As such, some of the generic
tests also required changes, as they did not include identifiers.

`nodes.json` was tested as a part of `nodes_test`.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>