Fix buffer size issue
`test-network-function` heavily relies upon the
`github.com/google/goexpect` go module as a dependency.  The module does a lot
of great things, but there are some faults.  Namely, `goexpect` abstracts a
bufferSize constant which is used for expect send/receive buffers.

https://github.com/google/goexpect/blob/master/expect.go#L34

On arbitrarily large payloads, this buffer is too small to suffice for certain
commands, particularly `oc get nodes -o json`.  The right long term solution is
to contribute a proper mechanism to adjust the buffer size in `goexpect`.
However, given the fact that we plan to release in 2 days, an alternative path
is proposed in this pull request.
1) github.com/google/goexpect is forked in github.com/ryandgoulding/goexpect.
   `goexpect` is licensed permissively as BSD3 and does not require copy-left.
   However, we still should contribute upstream to it eventually!
2) bufferSize is doubled.  This was tested in the New Mexico lab, and is
   sufficient for our client implementation.
3) Address a means of growing or setting buffer size upstream in `goexpect`
   eventually.  However, we should not gate our release on getting a PR in
   there.

This change is temporary, and will allow us to get what we need with minimal
cost and invasion.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>