Update test configuration and best practices
CTONET-626 is a story to solidify the way we handle test-dependent config
needed by test implementations.  Management of this configuration is important
since needs to be included in a claim file for reproducibility reasons.

This PR moves all of the various test configurations into the `pkg/config`
module.  Additionally, it abstracts a sane pattern for naming input test config
files in the manner `<configKey>_test_configuration`.  The extension/suffix is
left off as either JSON or YAML can be used.

The container/operator configuration was renamed to "cnf".  Although CNFs do not
need operators, CNFs have lifecycle management by operators.  If an operator is
not being tested, simply don't include it in the config.  Thus, we have two
configuration files:

* cnf_test_configuration.yml
* generic_test_configuration.yml

These two should be squashed into a single configuration at a later date, but
this sort of invasive change is out of scope this close to the release of
`1.0`.

All corresponding documentation was updated to reference the renamed files
appropriately.  Some additional documentation was added to "Test Configuration"
in README.md instructing developers that configuration necessary to reproduce
a claim *must* be added to the pool.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>