Fix a broken link to claim.schema.json
`claim.schema.json` is defined in
`github.com/redhat-nfvpe/test-network-function-claim`.  As part of the hectic
preparation to release `v1.0.0`, the location of `claim.schema.json` was moved
to the `schemas` directory as a part of #15:

https://github.com/redhat-nfvpe/test-network-function-claim/pull/15

This change adapts `test-network-function` to utilize the new location of the
`claim.schema.json` file.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>