Remove unneeded "GO111MODULE"
GO111MODULE was added in Go 1.11 to allow a developer to
force the (then) new behaviour of using Go Modules, vendoring,
and being able to specify specific versions. In 1.11 and 1.12
the new behaviour would not be enabled by default when a
project was on the configured GOPATH, meaning this option was
widely used.

In Go 1.13 and greater the default behaviour changed; the Go
Modules approach became the default ANY time a `go.mod` was
present, even when on the GOPATH.

Since we require Go 1.14 or greater and have a 'go.mod' file
the behaviour triggered by  "GO111MODULE=on" is the
default behaviour for our project, so it does not need to be
specified.

For further information and some adjacent issues see
https://dev.to/maelvls/why-is-go111module-everywhere-and-everything-about-go-modules-24k

Signed-off-by: Charlie Wheeler-Robinson <cwheeler@redhat.com>