refresh test result constants
The constants that define test results were defined as
`iota` values, in alphabetical order. This means that a
result of `0` was considered to be an error. The ordering is
very likely the result of a linting rule being applied
overzelously.

The use of both `tnf.ERROR` and other results as the value
for calls to `os.Exit` strongly implies that the `SUCCESS`,
i.e. test pass, should be `0`, indicating a successful execution
and allowing for use in CI pipelines.

This change addresses the issue by individually defining the
constants and updating documentation, comments, unit tests, and
JSON-defined CNF tests to use the new values.

Signed-off-by: Charlie Wheeler-Robinson <cwheeler@redhat.com>