Ease the pain of integrating new PTY executables
Google's goexpect provides means of spawning PTYs for generic shell
interfaces.  As such, CNF Certification 1.0.0 provided custom implementations
for oc, ssh, and shell PTYs.  As a result of working with various partners, it
became clear that we would eventually need the ability to spawn PTYs, and the
configuration of such PTYs should be made as painless as possible.

This PR introduces the ability to spawn a generic PTY based on a PTY
configuration YAML file.  Additionally, this file may be templated (i.e.,
provide dynamic arguments to a given command) using Go template language.  This
is particularly useful for commands such as "ftp", which may require a dynamic
destination address.  Configuration is handled using key/value pairs.

The "jsontest-cli" command was modified to include the ability to run in "pty"
and "pty-template" contexts.  This was done to support rapid development of new
PTY implementations, which can be played with and controlled directly from the
command line.

This PR boasts unit test coverage for all lines possible.  Certain lines of code
are unreachable, as the conditionals evaluated are false tautologies (i.e.,
given the implementation, they will never be executed).  The conditionals are
defensive in case the API or implementation changes in the future, which could
have the possibility of exercising the conditions.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>