Update to the latest version of google/goexpect
For test-network-function 1.0, we encountered an issue where Expecter output
would regularly overrun the buffers.  Because the buffer byte size was not
configurable, we forked the project to increase the buffer size as a stop gap
measure before including a proper fix.

The proper fix involved adding in a configurable bufferSize per GExpect
instance:

https://github.com/google/goexpect/pull/65

Since #65 has been merged, we can now remove our fork and use the proper
upstream offering of google/goexpect.

Additionally, this patch provides a Facade over expect.SpawnGeneric's
expect.Option optionals.  Essentially, we want to mandate a larger default
buffer size for test-network-function.  However, this introduces a conundrum:
there is no possible way to inspect whether a client has included
expect.BufferSize(...) in its expect.Options.  Reflection is not an option,
as expect.Option(s) return function pointers, which are not identifiable or
distinct in any way.  Thus, a facade was introduced by which the most popular
expect.Option(s) are front-ended by interactive.Option(s).  This provided us
the ability to track state, and to determine whether the larger default should
be included (i.e., the client did not specify expect.BufferSize(...)).

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>