Add configuration and documentation surrounding buffer sizes
Buffer sizes cannot be made variable.  This is an unfortunate limitation of
the PTY library used by google/goexpect.  The issue is rooted in the fact that
a PTY is treated like a file that is not meant to close (receive io.EOF)
between commands.

As such, a configuration knob was added to allow setting interactive.BufferSize
for each interactive context.  If a client does not supply a BufferSize, then
the default is determined.

The default is determined by attempting to read the TNF_DEFAULT_BUFFER_SIZE
environment variable, and then parsing the result as an int.  If the
environment variable is not set or cannot be parsed as an int, then a const
default is used (16KB).

Additionally, this PR adds documentation to the README explaining the situation
in which the default buffer size may need to be set.  An example of setting the
environment variable and then running run-cnf-suites.sh is provided.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>