Fix problems with the linter workflow
The linter workflow continues to be plagued by problems.  Namely, `make lint`
was not properly failing for `golint` errors.  In order to fix this, the
`lint` Makefile target was modified to invoke `golint` with the
`-set_exit_status` flag, which will exit 1 if errors are found.

Secondly, in order to avoid any ambiguity, we specify the checkout ref as
${{ github.sha }}.  This ensures that the branch or tag ref that triggered the
workflow run is used as the ref during checkout.  This ensures identical
behavior for both "push" and "pull" workflow invocations.

Lastly, existing lint issues are fixed so the build passes.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>