Combine the config mechanisms
Previously the "generic" and "multus" specs were in one config file
and "container" and "operator" specs were in a different file.

This change combines the two into a single config file that is
loaded once and then available to any spec on-demand. The location
of this config file can be set with an environment variable.

This also means that the configpool is no longer needed as a
seperate item.

As part of this change set the complex keys from the "generic"
config file were removed to improve compatibility with JSON. The
data is now in a list of maps. The removal of this natural key
does not impact performance as the first thing done after the
"generic" config is loaded is a new map is built.

Signed-off-by: Charlie Wheeler-Robinson <cwheeler@redhat.com>