WiP working config example for the generic spec
New config mechanism works as both the config pool and
the single point of config discovery and loading and
provides those to the specs.

Have not used https://github.com/tliron/yamlkeys as the
requirement to use custom map methods is a step too far
given the control over the source data in this case.

Still to do:
- convert `cnf` config
- fold in any needed config pool capabilities
- remove defunct config package code
- support json config files for parity with prev. implementation
- include raw config in claim?
- update docs: be sure to explain new restrictions due to no custom marshal/unmarshal available on the raw config side.
  - Comes down to must unmarshal into `map[string]interface`.
- more and better tests

Signed-off-by: Charlie Wheeler-Robinson <cwheeler@redhat.com>