CTONET-645 - scc verification
In this test we check whether `clusterPermissions` are set for specific
`ClusterServiceVersion` is installed. We test etcd operator which is not
using the permissions but this tests enable us to check whether operator
in test is using `privileged` and in this case we should fail and
analyze why such permission is needed.

Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>