automatically load generic config
Move the individual configuration section definitions to a new package
to avoid circular imports with autoconfig.

Create the autoconfig package that will find pods from a label,
and can process those pods into a working config for the generic
spec. This automatic config is controlled by an environment
variable.

Updated the README to explain the behaviour of automatic config.
Moved the explanation of the configuration options to a new
`docs` folder in the expectation that the configuration docs
will expand and change significantly.

Signed-off-by: Charlie Wheeler-Robinson <cwheeler@redhat.com>