Add an optional parameter -i to run-container.sh
Allows the end user to choose a custom TNF image with an optional `-i`
parameter.

This makes using a locally built TNF image finally possible.
Additionally, the `-i` parameter can be used to run a different version
of the TNF image, or an image from another registry.

If the origin of the given image cannot be verified, a warning message
is shown to protect the end user from pulling a wrong image by mistake.
The implementation of the check is naive. It's designed to help with
debugging image name misspellings, it does not verify if the pulled
TNF image has been tampered with or not.

Signed-off-by: Marek Kochanowski <mkochanowski@redhat.com>