decouple certification request from cnf info
The previous combination was misleading as the request info could
be any arbitrary strings and not anything to do with the CNF. This
change moves the request info to the top level of the config file
to make this lack of relationship clear, and restructures the
container test spec slightly to better accomodate this change.

There is also a change to the tests for this aspect of the config
that could be applied to the rest of the container and operator
config tests to improve them.

Signed-off-by: Charlie Wheeler-Robinson <cwheeler@redhat.com>