CTONET-645 - fix empty resourceNames
In case when resourceNames is not defined we returned empty string
which failed the test. This PR checkes whether there are no resource
names and make sure we use well know token `EMPTY` to indicate this.
The tocken was introduced to help validate whether path to

.spec.install.spec.clusterPermissions[].rules[].resourceNames

is valid. In order to use command introduced by this test in a test
description we can use action deny with expectedStatus `privileged`.

Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>