Add support for external TNF source repositories
This change adds an optional build argument `TNF_SRC_URL` to the
Dockerfile that allows to override the URL the TNF framework will be
downloaded from during the image build process. The value of the build
argument points to the official TNF repository by default.

Also, the Dockerfile now supports checking out a commit of a TNF source
repository, in addition to branches and tags.

Signed-off-by: Marek Kochanowski <mkochanowski@redhat.com>