Add network mode settings to run-container.sh
This change adds an optional -n parameter to run-container.sh allowing
to set network settings consumed by a `--network` parameter of the
`docker run` command responsible for creating the TNF container.
The network mode defaults to `bridge`, consistent with the equivalent
setting provided by the `docker run` command.

The documentation, and the usage prompt have been updated to explain
the usage of the -n parameter.

Due to the increase in the number of options available, the list of
optional arguments in the documentation is now sorted alphabetically.

Signed-off-by: Marek Kochanowski <mkochanowski@redhat.com>