add a note about certification status async failures
This was noticed when the following error occurred, claiming the
container was not certified because a section of the configuration was
empty:
```
   container rhel8/nginx-116 should eventually be verified as certified [It]
    /Users/cwheeler/cnf-cert/test-network-function/test-network-function/container/suite.go:76

    Expected
        <[]configsections.Cnf | len:0, cap:0>: nil
    not to be nil

    /Users/cwheeler/cnf-cert/test-network-function/test-network-function/container/suite.go:89
```

The test context (`[It]`) is not matched with the error, and the line numbers do not agree. The actual
line this error occurred on was `89`, the reference to `76` is for the previous line.

Signed-off-by: Charlie Wheeler-Robinson <cwheeler@redhat.com>