clarify misleading wording of container spec config
The entire spec is called the `container` spec, but is actually
testing Pods and the more-nebulous CNFs.
This misleading nomenclature is present throughout the code and
needs a much more wide-ranging change to address it longer-term.

This commit changes the comments describing the configuration
file to make it slightly clearer to a user exactly what they
should add to the file.

Signed-off-by: Charlie Wheeler-Robinson <cwheeler@redhat.com>