correct error and behaviour in CNF autodiscovery
The error was for a `generic` case, and the attempt to continue
would always lead to a later and less well-explained failure in
the `container` spec. This error halts the whole test run, which
is deemed to be acceptable in the case of  a mismatch between
configuration and invocation instructions to the app.

Signed-off-by: Charlie Wheeler-Robinson <cwheeler@redhat.com>