Explicitly allow use of "docker" instead of "podman"
Resolves #238.

There are a whole host of issues with using "podman" on many non-linux
Operating Systems.  For example, support in macOS currently involves federation
with a remote linux based server.  As such, we should support multiple
platforms and allow the use of alternative container virtualization clients.

This change adds the auto-detection of a user-specified container
virtualization client through the TNF_CONTAINER_CLIENT environment variable.
If the user does not set the variable, then "podman" is used by default.

Additionally, documentation was added to give an example of how to use
"docker" instead of "podman".

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>