Add in functionality to uncordon nodes
Uncordon nodes that are drained so that they are not left idle.
This code is unit tested.  The test is implemented using a JSON
template, and a new API was added to utilize a map to fill in the
template values.

A few identifiers were fixed as a part of this commit.

Finally, TNF_NON_INTRUSIVE_ONLY was added to only run Non-intrusive
test specs.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>