Resolves issues with checking operator subscription
Hitherto, the operator subscription is tested against a non-templated
test case input file.  The existing framework does not have the innate
ability to render the file contents based on a template, making it an
unsuitable candidate for a test which requires specific, templated output.
Since test cases ultimately are fed into the tool using a Go string,
templating this code seemed like the wrong solution.

This change rips the "CHECK_SUBSCRIPTION" test out of the existing operator
testing framework.  That framework is difficult to maneuver, and this change
aligns with utilizing the JSON test framework instead.  Why was this approach
chosen?  Here are some of the reasons:

1) Standardize on JSON Test Template framework.
2) Disaggregate the reference to the "operator-is-installed-via-olm" from the
   existing "operator-best-practices" test case.  This provides further
   granularity so that this test case can pass/fail independently of the other
   operator best practices.
3) Unit testing is crafted towards the specific use case, instead of testing
   pseudo-template rendering.  In this case, we can use actual input/output
   that the Handler would encounter.
4) The code is easier to read.

Finally, CATALOG.md is regenerated to reflect the changes.  This involved
eliminating the "pod-is-recreated" test case from the catalog, which was
removed in an earlier PR.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>